requirements.txt
Django == 3.1.1
wheel == 0.35.1
coreapi == 2.3.3
Markdown ==  3.2.2
Pygments == 2.6.1
django-filter == 2.3.0
djangorestframework == 3.11.1
drf-yasg == 1.17.1
python-dotenv == 0.14.0
six~=1.15.0